finalizing communication with the API

+++ src/components/Demo.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect } from "react"
+import React, { useState, useEffect } from "react"
 
 import{copy, linkIcon, loader, tick} from '../assets'
 import {useLazyGetSummaryQuery} from '../services/article'
@@ -15,6 +15,9 @@ const Demo = () => {
   const [getSummary, {error, isFetching}] = 
   useLazyGetSummaryQuery();
 
+  const [copied, setCopied] = useState("");
+  
+
   useEffect(() => {
     const articlesFromLocalStorage = JSON.parse(
       localStorage.getItem('articles')
@@ -28,17 +31,35 @@ const Demo = () => {
   const handleSubmit = async (e) => {
     // API request
     e.preventDefault();
+    const existingArticle = allArticles.find(
+      (item) => item.url === article.url
+    );
+
+    if (existingArticle) return setArticle(existingArticle);
+
     const {data} = await getSummary({articleUrl: article.url});
 
     if (data?.summary) {
       const newArticle = {...article, summary: data.summary};
 
-      const updatedArticles = [newArticle, ...allArticles];
+      const updatedAllArticles = [newArticle, ...allArticles];
 
       setArticle(newArticle);
-      setAllArticles(updatedArticles);
+      setAllArticles(updatedAllArticles);
 
       localStorage.setItem('articles', JSON.stringify(updatedAllArticles));
+    }
+  }
+
+  const handleCopy = (copyUrl) => {
+    setCopied(copyUrl);
+    navigator.clipboard.writeText(copyUrl);
+    setTimeout(() => setCopied(false), 3000);
+  }
+
+  const handleKeyDown = (e) => {
+    if (e.key === 13) {
+      handleSubmit(e);
     }
   }
 
@@ -63,6 +84,7 @@ const Demo = () => {
             value={article.url}
             onChange={(e) => {setArticle({...
               article, url: e.target.value})}}
+            onKeyDown={handleKeyDown}  
             required
             className="url_input peer"
           />
@@ -79,16 +101,17 @@ const Demo = () => {
 
         {/* Browser URL History */}
         <div className="flex flex-col gap-1m max-h.60 overflow-y-auto">
-          {allArticles.map((item,index) => (
+          {allArticles.reverse().map((item,index) => (
             <div
               key={`link-${index}`}
               onClick={() => setArticle(item)}
               className="link_card"
               >
-                <div className="copy_btn">
+                <div className="copy_btn"
+                onClick={() => handleCopy(item.url)}>
                  <img 
-                 src={copy}
-                 alt="copy_icon"
+                 src={copied === item.url ? tick : copy}
+                 alt={copied === item.url ? 'ticked_icon' : 'copy_icon'}
                  className="w-[40%] h-[40%] object-contain"/> 
                 </div>
                 <p className="flex-1 font-satoshi text-blue-700 font-medium text-sm truncate">
